+++ src/transforms/minify.ts
@@ -15,6 +15,7 @@ import {
   AssignmentExpression,
   VariableDeclarator,
   Identifier,
+  CallExpression,
 } from "../util/gen";
 import {
   getBlockBody,
@@ -22,11 +23,14 @@ import {
   isForInitialize,
   isLexContext,
   getFunction,
+  prepend,
+  append,
 } from "../util/insert";
 import { isValidIdentifier, isEquivalent } from "../util/compare";
 import { walk, isBlock } from "../traverse";
 import { ok } from "assert";
 import { isLexicalScope } from "../util/scope";
+import Template from "../templates/template";
 
 /**
  * Basic transformations to reduce code size.
@@ -37,12 +41,13 @@ import { isLexicalScope } from "../util/scope";
  * - `x['y']` **->** `x.y`
  */
 export default class Minify extends Transform {
-  variables: Map<Node, Location[]>;
+  /**
+   * A helper function that is introduced preserve function semantics
+   */
+  arrowFunctionName: string;
 
   constructor(o) {
     super(o, ObfuscateOrder.Minify);
-
-    this.variables = new Map();
   }
 
   match(object: Node, parents: Node[]) {
@@ -239,17 +244,42 @@ export default class Minify extends Transform {
         });
 
         if (canTransform) {
+          if (!this.arrowFunctionName) {
+            this.arrowFunctionName = this.getPlaceholder();
+
+            append(
+              parents[parents.length - 1] || object,
+              Template(`
+            function ${this.arrowFunctionName}(arrowFn){
+              return function(){ return arrowFn(...arguments) }
+            }
+            `).single()
+            );
+          }
+
+          const wrap = (object: Node) => {
+            return CallExpression(Identifier(this.arrowFunctionName), [
+              clone(object),
+            ]);
+          };
+
           if (object.type == "FunctionExpression") {
             object.type = "ArrowFunctionExpression";
+
+            this.replace(object, wrap(clone(object)));
           } else {
             var arrow = { ...clone(object), type: "ArrowFunctionExpression" };
             this.replace(
               object,
-              VariableDeclaration(VariableDeclarator(object.id.name, arrow))
+              VariableDeclaration(
+                VariableDeclarator(object.id.name, wrap(arrow))
+              )
             );
 
             var x = this.transform(arrow, []);
-            x();
+            if (typeof x === "function") {
+              x();
+            }
           }
         }
       };
@@ -442,6 +472,8 @@ export default class Minify extends Transform {
         }
 
         if (
+          object.consequent &&
+          object.consequent.body &&
           object.consequent.body.length == 1 &&
           object.alternate &&
           object.alternate.body.length == 1